requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
pydantic
pyyaml
httpx
sentence-transformers
